refactor - rename variable, extract method, move method, pull up method, extract class
