+++ src/HackerRank/src/Tracks/Tutorials/ThirtyDaysOfCode/D22_BinarySearchTrees/Java/Solution.java
@@ -5,51 +5,17 @@ import java.util.*;
 public class Solution {
 
     public static void main(String[] args) {
+        Node node = new Node();
         Scanner sc=new Scanner(System.in);
         int T=sc.nextInt();
         Node root=null;
         while(T-->0){
             int data=sc.nextInt();
-            root=insert(root,data);
+            root=node.insert(root,data);
         }
-        int height=getHeight(root);
+        int height=node.getHeight(root);
         System.out.println(height);
         sc.close();
     }
-
-    public static Node insert(Node root,int data){
-        if(root==null){
-            return new Node(data);
-        }
-        else{
-            Node cur;
-            if(data<=root.data){
-                cur=insert(root.left,data);
-                root.left=cur;
-            }
-            else{
-                cur=insert(root.right,data);
-                root.right=cur;
-            }
-            return root;
-        }
-    }
-
-    public static int getHeight(Node root){
-        if (root == null){
-            return -1;
-        }
-        else {
-            return 1+ Math.max(getHeight(root.left), getHeight(root.right));
-        }
-    }
 }
 
-class Node{
-    Node left,right;
-    int data;
-    Node(int data){
-        this.data=data;
-        left=right=null;
-    }
-}
