+++ src/LeetCode/src/SwapNumbersInArray.java
@@ -1,0 +1,9 @@
+package LeetCode.src;
+
+public class SwapNumbersInArray {
+    public static void swap(int[] nums, int i, int j) {
+        int temp = nums[i];
+        nums[i] = nums[j];
+        nums[j] = temp;
+    }
+}
